Fix index name collisions on schema re-creation
